backend/requirements.txt
Django==1.11.13
djangorestframework==3.8.2
MySQL-python==1.2.5
pytz==2018.4
